upgrade query params extension to new template
