+++ jlab_query_params/.prettierignore
@@ -2,4 +2,5 @@ node_modules
 **/node_modules
 **/lib
 **/package.json
-shiny_proxy_jlab_query_parms
+!/package.json
+jlab_query_params
