+++ jlab_query_params/.yarnrc.yml
@@ -1,0 +1,3 @@
+enableImmutableInstalls: false
+
+nodeLinker: node-modules
