+++ jlab_query_params/install.json
@@ -1,0 +1,5 @@
+{
+  "packageManager": "python",
+  "packageName": "jlab_query_params",
+  "uninstallInstructions": "Use your Python package manager (pip, conda, etc.) to uninstall the package jlab_query_params"
+}
