+++ jlab_query_params/jlab_query_params/handlers.py
@@ -2,14 +2,12 @@ import json
 import os
 import logging
 from urllib.parse import parse_qsl
+
 from jupyter_server.base.handlers import APIHandler
 from jupyter_server.utils import url_path_join
 import tornado
 
 class RouteHandler(APIHandler):
-    # The following decorator should be present on all verb methods (head, get, post,
-    # patch, put, delete, options) to ensure only authorized user can request the
-    # Jupyter server
     @tornado.web.authenticated
     def post(self):
         body = self.request.body.decode('utf-8')
@@ -25,6 +23,6 @@ def setup_handlers(web_app):
     host_pattern = ".*$"
 
     base_url = web_app.settings["base_url"]
-    route_pattern = url_path_join(base_url, "shiny-proxy-jlab-query-parms", "set_env_vars")
+    route_pattern = url_path_join(base_url, "jlab-query-params", "set-env-vars")
     handlers = [(route_pattern, RouteHandler)]
     web_app.add_handlers(host_pattern, handlers)
