+++ jlab_query_params/jupyter-config/server-config/jlab_query_params.json
@@ -1,7 +1,7 @@
 {
   "ServerApp": {
     "jpserver_extensions": {
-      "shiny_proxy_jlab_query_parms": true
+      "jlab_query_params": true
     }
   }
 }
