+++ jlab_query_params/style/index.css
@@ -1,2 +1,1 @@
-
 @import url('base.css');
