--- shiny_proxy_jlab_query_parms/.eslintignore
@@ -1,5 +1,0 @@
-node_modules
-dist
-coverage
-**/*.d.ts
-tests
