--- shiny_proxy_jlab_query_parms/.prettierrc
@@ -1,5 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "none",
-  "arrowParens": "avoid"
-}
