--- shiny_proxy_jlab_query_parms/MANIFEST.in
@@ -1,24 +1,0 @@
-include *.md
-include pyproject.toml
-recursive-include jupyter-config *.json
-
-include package.json
-include install.json
-include ts*.json
-include yarn.lock
-
-graft shiny_proxy_jlab_query_parms/labextension
-
-# Javascript files
-graft src
-graft style
-prune **/node_modules
-prune lib
-prune binder
-
-# Patterns to exclude from any directory
-global-exclude *~
-global-exclude *.pyc
-global-exclude *.pyo
-global-exclude .git
-global-exclude .ipynb_checkpoints
