--- shiny_proxy_jlab_query_parms/pyproject.toml
@@ -1,17 +1,0 @@
-[build-system]
-requires = ["jupyter_packaging~=0.10,<2", "jupyterlab~=3.1"]
-build-backend = "jupyter_packaging.build_api"
-
-[tool.jupyter-packaging.options]
-skip-if-exists = ["shiny_proxy_jlab_query_parms/labextension/static/style.js"]
-ensured-targets = ["shiny_proxy_jlab_query_parms/labextension/static/style.js", "shiny_proxy_jlab_query_parms/labextension/package.json"]
-
-[tool.jupyter-packaging.builder]
-factory = "jupyter_packaging.npm_builder"
-
-[tool.jupyter-packaging.build-args]
-build_cmd = "build:prod"
-npm = ["jlpm"]
-
-[tool.check-manifest]
-ignore = ["shiny_proxy_jlab_query_parms/labextension/**", "yarn.lock", ".*", "package-lock.json"]
